adjusted code

+++ src/functions/delete_product.rs
@@ -5,12 +5,16 @@ use crate::functions::all_errors::{MyError};
 
 pub(crate) async fn delete() -> Result<(), MyError> {
     let term = Term::stdout();
-    
+    loop {
         let product_id: i32 = Input::new()
             .with_prompt("Enter product ID number")
             .interact_text()?;
 
         
+        if product_id == 0 {
+            break;
+        }
+
         let pool = SqlitePool::connect("./Database/prod.db").await?;
 
         let result = sqlx::query(" DELETE FROM products WHERE id = ?")
@@ -30,7 +34,8 @@ pub(crate) async fn delete() -> Result<(), MyError> {
             .interact()?;
 
         term.clear_last_lines(100);
-        
+
+     
+    }
     Ok(())
-    
 }
